requirementes.txt
aniso8601==9.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
google-auth==2.3.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
